+++ ext/node/lib.rs
@@ -713,7 +713,7 @@ deno_core::extension!(deno_node,
     "node:path" = "path.ts",
     "node:path/posix" = "path/posix.ts",
     "node:path/win32" = "path/win32.ts",
-    "node:perf_hooks" = "perf_hooks.ts",
+    "node:perf_hooks" = "perf_hooks.js",
     "node:process" = "process.ts",
     "node:punycode" = "punycode.ts",
     "node:querystring" = "querystring.js",
